GridMap can safely contain Array instead of...
OrderedSet values. An OrderedSet here contains an Array already, and the only reason for it to also have a Set part is to ensure uniqueness. However, we add every DSGItem in the Array anyway to another OrderedSet, which enforces uniqueness there anyway. This seems to gain a tiny bit of FPS at larger sizes (over 8000 balls).