Clamping velocity seems to work...?
The clamp is component-wise, but could be adjusted a lot. it depends on the current delta, but inverted so larger delta (at slower framerates) are clamped to slightly tighter bounds. I don't know how right or wrong this is, but it seems to prevent the massive chaotic movements with larger amounts of balls.

This also uses the latest changes in Solver suggested on Discord, but not committed yet. First!